Another bug fix in attach_uvw_uvfits

+++ src/craco/attach_uvws_uvfits.py
@@ -50,7 +50,7 @@ def main():
         fast_blocker = f.uvsource.fast_time_blocks(nt = end_samp - args.start_samp + 1, fetch_uvws=True)
         data, uvws = next(fast_blocker)
         uvws = convert_uvws_list_to_array(uvws)
-        f.swap_with_data(data[:, 0, 0, 0], bl2array(uvws))
+        f.swap_with_data(data[:, 0, 0, 0], bl2array(uvws, dtype=np.float64))
     elif args.use_visrows:
         print("Using visrows()")
         pass
